refactor : move all `constants`related files into `index.ts`
