+++ src/components/common/CustomSelect.tsx
@@ -4,9 +4,7 @@ import { FieldError } from 'react-hook-form';
 import { BiSolidChevronRight } from 'react-icons/bi';
 import { Button } from '.';
 import type { ExpenseTracker, RestrictedRecipeForValidation } from '../../supabase/schema';
-import { customPropReceiver } from '../../constants';
-import { PaymentDataType } from '../../constants/expenseTracker';
-import { FilmRecipeFieldDataType } from '../../constants/recipes';
+import { customPropReceiver, PaymentDataType, FilmRecipeFieldDataType } from '../../constants';
 
 export type CustomSelectDataType = PaymentDataType | FilmRecipeFieldDataType[number] | string | number;
