+++ src/components/expenseTracker/TotalExpensePrice.tsx
@@ -1,10 +1,8 @@
 import styled from '@emotion/styled';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getAllPaymentsByMonth } from '../../supabase/expenseTracker';
-import queryKey from '../../constants/queryKey';
-import { MatchedPriceUnitWithSymbol, matchedPriceUnitWithSymbol } from '../../constants/expenseTracker';
+import { MatchedPriceUnitWithSymbol, matchedPriceUnitWithSymbol, staleTime, queryKey } from '../../constants';
 import { monetizeWithSeparator } from '../../utils/money';
-import { staleTime } from '../../constants/staleTime';
 
 interface TotalExpensePriceProps {
 	currentMonthIndex: number;
