+++ src/components/todoReminder/TodoItem.tsx
@@ -10,8 +10,7 @@ import { Button, Checkbox, TextInput } from '../common';
 import { useDrag, useLoading } from '../../hooks';
 import { useToastStore, useModalStore } from '../../store';
 import { editTodoContent, removeTodo, updatedTodoCompleted } from '../../supabase/todos';
-import queryKey from '../../constants/queryKey';
-import { toastData } from '../../constants/toast';
+import { toastData, queryKey } from '../../constants';
 import { today } from '../../utils/date';
 import { MODAL_CONFIG } from '../modal/modalType';
 import { todoItemSchema, TodoItemSchema } from './schema';
