+++ src/constants/routes.ts
@@ -14,5 +14,4 @@ const routes = {
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type Route<T> = T[keyof T] | (string & {});
-
-export default routes;
+export { routes };
