+++ src/hooks/mutations/useAddDiaryMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { addDiary } from '../../supabase/diary';
 import { Diary } from '../../supabase/schema';
-import queryKey from '../../constants/queryKey';
+import { queryKey } from '../../constants';
 
 const add = (data: Diary) => (oldData: Diary[]) => {
 	return [...oldData, data];
