+++ src/hooks/mutations/useRemoveDiaryMutation.ts
@@ -3,9 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { removeDiary } from '../../supabase/diary';
 import { Diary } from '../../supabase/schema';
 import { useToastStore } from '../../store';
-import queryKey from '../../constants/queryKey';
-import { routes } from '../../constants';
-import { toastData } from '../../constants/toast';
+import { toastData, routes, queryKey } from '../../constants';
 
 const remove =
 	({ id }: { id: string }) =>
