+++ src/pages/ExpenseTrackerFixedCost.tsx
@@ -1,12 +1,10 @@
 import styled from '@emotion/styled';
 import { useNavigate } from 'react-router-dom';
 import { useSuspenseQuery } from '@tanstack/react-query';
-import queryKey from '../constants/queryKey';
+import { queryKey, bankSvgs, routes } from '../constants';
 import { getFixedCostPaymentsByMonth } from '../supabase/expenseTracker';
 import { getDateFromString, getNextMonthFormatDate, months, today } from '../utils/date';
-import { bankSvgs } from '../constants/expenseTracker';
 import { monetizeWithSeparator } from '../utils/money';
-import { routes } from '../constants';
 
 /**
  * 고정 지출 내역
